// src/api/models/analyticsSummaryResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Analytics Service API
 * Analytics Service API
 * OpenAPI spec version: 1.0
 */
import type { CategoryAnalyticsResponse } from './categoryAnalyticsResponse';
import type { UpcomingChargesResponse } from './upcomingChargesResponse';

/**
 * Сводная аналитика по подпискам пользователя
 */
export interface AnalyticsSummaryResponse {
  /** Общее количество подписок */
  totalSubscriptions?: number;
  /** Суммарная стоимость всех подписок */
  totalAmount?: number;
  /** Статистика по категориям */
  categories?: CategoryAnalyticsResponse[];
  /** Количество ближайших списаний */
  upcomingCharges?: UpcomingChargesResponse;
}