// src/api/models/categoryAnalyticsResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Analytics Service API
 * Analytics Service API
 * OpenAPI spec version: 1.0
 */
import type { CategoryAnalyticsResponseCategory } from './categoryAnalyticsResponseCategory';

/**
 * Разбивка статистики по категории
 */
export interface CategoryAnalyticsResponse {
  /** Категория подписок */
  category?: CategoryAnalyticsResponseCategory;
  /** Количество подписок в категории */
  subscriptionCount?: number;
  /** Суммарная стоимость подписок в категории */
  totalAmount?: number;
}